claude-agent-tutorial/pyproject.toml
[project]
name = "cc-agent-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.71.0",
    "claude-agent-sdk>=0.1.4",
    "e2b>=2.3.2",
    "e2b-code-interpreter>=2.2.0",
    "google-genai>=0.5.0",
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "mcp>=1.18.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pypdf>=6.2.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "rich>=13.0.0",
]

[tool.uv]
managed = true
